requirements.txt
# AI-Horizon Ed Dependencies

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
# sqlite3 is built-in with Python
psycopg2-binary==2.9.7  # For PostgreSQL connection (Heroku)

# HTTP Requests
requests==2.31.0

# YouTube Data API
google-api-python-client==2.100.0

# GitHub API
PyGithub==1.59.1

# Course Platform APIs
beautifulsoup4==4.12.2
lxml==4.9.3

# Date/Time Handling
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Logging
colorama==0.4.6

# JSON Processing
simplejson==3.19.1

# URL Parsing
urllib3==2.0.4

# AI APIs for Resource Discovery and Scoring
anthropic>=0.3.0
openai>=1.0.0

# Async Support
aiohttp==3.8.5

# Production Server (Required for Heroku)
gunicorn==21.2.0

# Optional: If integrating with main platform database
# psycopg2-binary==2.9.7  # For PostgreSQL connection 